% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_count_trips.R
\name{aa_count_trips}
\alias{aa_count_trips}
\title{Basic stats from Trip AA tables and ADIOS}
\usage{
aa_count_trips(channel, yearRange = 1964:2017)
}
\arguments{
\item{channel}{an RODBC object (see \code{\link{connect_to_database}})}

\item{yearRange}{a vector or years. Default =1964:2017}
}
\value{
A matrix 54 x 3

\code{tableCount}{

Column 1 - The year

Column 2 - The number of Trips from the AA tables

Column 3 - The number of trips from ADIOs (mv_cf_landings)

Column 4 - Difference in number of Trips
}
}
\description{
Exploratory code to examine difference in number of unique trips between AA tables and ADIOS (mv_cf_landings).
}
