% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gear.R
\name{aggregate_gear}
\alias{aggregate_gear}
\title{Aggregates landings and length data based on gear}
\usage{
aggregate_gear(data, recodeOtherGear, landingsThresholdGear)
}
\arguments{
\item{data}{List. Landings data and length data}

\item{recodeOtherGear}{Numeric scalar. Arbitrary code to use for "other gears" that contribute little to landings}

\item{landingsThresholdGear}{Numeric scalar. (Proportion). Determins the proportion of landings that have to be satisfied before all other landings by gear are lumped into "other gear" category}
}
\value{
List

\item{landings}{same as input}
\item{lengthData}{Same as input}
}
\description{
Aggregates landings and length data based on gear
}
