% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_landings.R
\name{aggregate_landings}
\alias{aggregate_landings}
\title{Landings data aggregated based on length sample availability}
\usage{
aggregate_landings(
  landingsData,
  lengthData,
  species_itis,
  landingsThresholdGear = 0.9,
  nLengthSamples = 1,
  pValue = 0.05,
  aggregate_to = NULL,
  borrowLengths = T,
  proportionMissing = 0.2,
  otherGear = "998",
  outputDir = here::here("output"),
  outputPlots = F,
  logfile = "logFile.txt"
)
}
\arguments{
\item{landingsData}{Tidy data frame. Landings by YEAR,QTR,NEGEAR,MARKET_CODE,landings_land,landings_nn,len_totalNumLen,len_numLengthSampls}

\item{lengthData}{Tidy data frame. Length data by YEAR,QTR,NEGEAR,MARKET_CODE, LENGTH, NUMLEN}

\item{species_itis}{Numeric scalar. species_itis code relating to the landings and length data}

\item{landingsThresholdGear}{Numeric scalar (proportion). Minimum proportion of cumulative landings to avoid aggregation of gear. Default = .9}

\item{nLengthSamples}{Numeric scalar. The minimum number of length sample sizes required to avoid combination of data. Default = 1}

\item{pValue}{Numeric scalar. Threshold pvalue for determining significance of ks test for length samples}

\item{aggregate_to}{Character string. Level of aggregation for all MARKET_CODES and gears (NULL, "QTR", "YEAR", "MIX").
Default = NULL - do not aggregate.}

\item{borrowLengths}{Boolean. Return data as is or use allgorithm to borrow lengths from neighboring cells (Time and/or space)}

\item{proportionMissing}{Numeric scalar. Proportion of missing samples allowed per YEAR for each MARKET_CODE/GEAR combination). Default = 0.2}

\item{otherGear}{Character string. Code to indicate the class for "other Gear".
This is the group of gear types that land the species of interest but in small numbers Default = "998"}

\item{outputDir}{Character string. Path to output directory (png files saved here)}

\item{outputPlots}{Boolean. Should plots be created. T or F (Default = F)}

\item{logFile}{Character string. Specify the name for the log file generated describing all decisions made.}
}
\value{
List of landings and associated length samples
\item{landings}{Tibble (n x 8). Aggregated landings data. YEAR, QTR, NEGEAR, MARKET_CODE,landings_land (metric tonnes), landings_nn (# trips), len_totalNumLen (# fish lengths), len_numLengthSamples (# independent samples) }
\item{lengthData}{Tibble (m x 8 ). Aggregated length data. YEAR, QTR, NEGEAR, MARKET_CODE, LENGTH (length of fish), NUMLEN (# fish at LENGTH)}
}
\description{
Landings data analysed to determine NEGEAR aggregation, MARKET_CODE descriptions analysed over time to determine aggregation all subject to
length sample availability. Output document created and a logfile to inform the user what steps were taken during aggregation
}
\details{
All plotting functions will need to be finined/generalized
rmd file needs to be created to report all decisions
}
