% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_market_codes.R
\name{aggregate_market_codes}
\alias{aggregate_market_codes}
\title{Aggregates landings and length data based on gear}
\usage{
aggregate_market_codes(data, pValue, outputDir, outputPlots, logfile)
}
\arguments{
\item{data}{List. Landings data and length data}

\item{pValue}{numeric scalar. Threshold pvalue for determining significance of ks test for length samples}

\item{outputDir}{Character string. Path to output directory (png files saved here)}

\item{outputPlots}{Boolean. Should plots be created. T or F (Default = F)}

\item{logFile}{character string. Specify the name for the log file generated describing all decisions made.}
}
\value{
List

\item{landings}{same as input}
\item{lengthData}{Same as input}
}
\description{
Aggregates landings and length data based on gear
}
