% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_qtr.R
\name{aggregate_to_qtr}
\alias{aggregate_to_qtr}
\title{Missing QTR length samples are filled using previous years QTR}
\usage{
aggregate_to_qtr(data, gearType, marketCode, QTRData, missingEarlyYears,
  nLengthSamples, pValue, outputDir, logfile)
}
\description{
Missing QTR length samples are filled using previous years QTR
}
