% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_year.R
\name{aggregate_to_year}
\alias{aggregate_to_year}
\title{Aggregate data to YEAR and borrow length samples from other years}
\usage{
aggregate_to_year(data, gearType, marketCode, aggYEARData, sampleStartYear,
  missingEarlyYears, proportionMissing, nLengthSamples, pValue, outputDir,
  logfile)
}
\description{
Data is too sparse at NEGEAR/MARKET_CODE level to retain at QTR year so it
is aggregated to YEAR level. YEARs with missing length samples are filled with length sample data from neighboring years
}
