% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_numbers_at_age.r
\name{calc_numbers_at_age}
\alias{calc_numbers_at_age}
\title{Calculates the numbers of fish at age}
\usage{
calc_numbers_at_age(expandedLandings, ageLenKeys, lengthWeightParams)
}
\arguments{
\item{expandedLandings}{Tibble. Expanded langings by length (from \code{expand_landings_to_lengths})}

\item{lengthWeightParams}{List. alpha = intercept, betas = slope(s), var = residual variance used to formulate the mean (?see Notes section below)}

\item{ageLengthKeys}{Tibble. Age Length Keys (from \code{create_yr_semester_age_length_key}}
}
\value{
A Tibble of expanded landings to represent weight of landings by length
\item{YEAR}{Year of landings}
\item{NEGEAR}{3 digit gear code as defined in cfdbs.gear}
\item{TIME}{Quarter/Half year of landings (The presence of this field depends on whether it was present in the \code{landingsData}}
\item{MARKET_CODE}{Market code assigned to landed fish (The presence of this field depends on whether it was present in the \code{landingsData}}
\item{LENGTH}{Length of sampled fish}
\item{NUMLEN}{number of sampled for fish stated LENGTH}
\item{weight}{expanded weight (mt) of all fish of given LENGTH in YEAR, NEGEAR etc..}
}
\description{
Given the landings (expanded by length), the age-length key and the length-weight relationship,
the number of fish at a given age is calculated
}
\section{Notes}{


The length weight relationship (see \code{\link{fit_length_weight}}) is fit assuming log normal errors (normal on the log scale).
Therefore when exponentiating a correction for the estimate is required:

E(W) = \eqn{\alpha  L^\beta exp(\sigma^2 / 2)}
}

