% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_length_distributions.R
\name{compare_length_distributions}
\alias{compare_length_distributions}
\title{Compare length distributions across categories}
\usage{
compare_length_distributions(landings, lengthData, variableToAggreagte,
  pValue, outputDir, logfile)
}
\arguments{
\item{landings}{tibble. landings data in form described below}

\item{lengthData}{tibble. length data in form described below}

\item{variableToAggreagte}{character string. Denote the variable to aggregate based on length distribution}

\item{pValue}{numeric scalar. Threshold pvalue for determining significance of ks test for length samples}

\item{outputDir}{Character string. Directory where output is saved}

\item{logfile}{Cahracter string. Name of logfile}
}
\value{
Updated landings and lengthsData object reflecting changes of aggregating categories over length
}
\description{
Compare length distributions across categories
}
