% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age_length_key.R
\name{create_age_length_key}
\alias{create_age_length_key}
\title{Creates Age length key}
\usage{
create_age_length_key(ageLengthData, aggregate_to)
}
\arguments{
\item{ageLengthData}{List. Landings data and length data}

\item{aggregate_to}{Character string. Level of aggregation for all MARKET_CODES and gears ("QTR", "YEAR", "SEMESTER", MIX").}
}
\value{
Data frame

\item{}{}
\item{}{}
}
\description{
Uses Gerritsen et al (2006) method for predicting missing values in
age length key
}
