% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filter_expression.R
\name{create_filter_expression}
\alias{create_filter_expression}
\title{create an expression to use in dplyr::filter statement}
\usage{
create_filter_expression(landingsRecord, categories)
}
\arguments{
\item{landingsRecord}{Tibble (1 x n). Single row from landingsData}

\item{categories}{Character vector. Names of Variables to filter over ("YEAR","QTR","NEGEAR", etc)}
}
\value{
An expression
}
\description{
create an expression to use in dplyr::filter statement
}
\section{Notes}{


Values are access from the \code{landingsRecord} using the \code{categories} and strung together to make an expression in the form of
labe == value & label = value etc.
}

