% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yr_semester_age_length_key.r
\name{create_yr_semester_age_length_key}
\alias{create_yr_semester_age_length_key}
\title{Creates Age length key}
\usage{
create_yr_semester_age_length_key(
  expandedLandings,
  ageLengthData,
  filterYrs,
  plusAge,
  referenceAge,
  printConvergence = F
)
}
\arguments{
\item{expandedLandings}{Data frame. Landings expanded by length distribution}

\item{ageLengthData}{Data frame. Age-length data from databases}

\item{filterYrs}{Numeric scalar. The First year in which age length key is required}

\item{plusAge}{Numeric scalar. Maximum age category. (eg , 10+)}

\item{referenceAge}{Numeric scalar.Used in the GLM as the age against which all
the other ages are estimated. It is typically the most abundant age.}

\item{printConvergence}{Boolean. Should convergence messages from \code{multinom} be printed to console (Default = F)}
}
\value{
Data frame

\item{}{}
\item{}{}
}
\description{
Uses Gerritsen et al (2006) method for predicting missing values in
age length key. Currently disregards gear type and only uses QTR/SEMESETER/YEAR.
}
