% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_landings_to_lengths.R
\name{expand_landings_to_lengths}
\alias{expand_landings_to_lengths}
\title{Proportions commercial landings based on length sample distribution}
\usage{
expand_landings_to_lengths(landingsData, lengthData, lengthWeightKey)
}
\arguments{
\item{landingsData}{Tibble. Aggregated landings data. YEAR, QTR, NEGEAR, MARKET_CODE,landings_land (metric tonnes), landings_nn (# trips), len_totalNumLen (# fish lengths), len_numLengthSamples (# independent samples).}

\item{lengthData}{Tibble. Aggregated length data. YEAR, QTR, NEGEAR, MARKET_CODE, LENGTH (length of fish), NUMLEN (# fish at LENGTH)}

\item{lengthWeightKey}{how define this? one over all time, time varying?}
}
\value{
Tibble
\item{expandedLandings}{landingsData expanded to represent weight of landings by length}
}
\description{
Proportions commercial landings based on length sample distribution
}
