% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_to_weight.R
\name{expand_to_weight}
\alias{expand_to_weight}
\title{Length expansion calculation}
\usage{
expand_to_weight(fishLength, numAtLength, landings, lengthWeightParams)
}
\arguments{
\item{fishLength}{Numeric vector. Lengths of fish in group}

\item{numAtLength}{Numeric vector. Number of fish at specified \code{fishLength}'s}

\item{landings}{Numeric vector. Landings total for group}

\item{lengthWeightParams}{List (3). List. alpha = intercept, betas = slope(s), var = residual variance used to formulate the mean}
}
\value{
vector
\item{fishWeight}{scaled landings for fish of specified \code{fishLength}}
}
\description{
Expand lengths in sample to weight and scale up to landings total.
}
\section{Notes}{


This an internal function called within a dplyr::mutate statement operating on a group
}

