% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_length_weight.R
\name{fit_length_weight}
\alias{fit_length_weight}
\title{Fit length weight relationship to species data from SVDBS}
\usage{
fit_length_weight()
}
\arguments{
\item{lengthWeight}{Data frame. length-weight pairs. Each row represents an individual fish}
}
\value{
Vector of parameters
\item{a}{}
\item{b}{}
}
\description{
Fits a length weight relationship for use in length expansion
}
