% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multinomial_props.R
\name{get_multinomial_props}
\alias{get_multinomial_props}
\title{Use multinomial function to automatically fill holes in age data}
\usage{
get_multinomial_props(
  agedata,
  small.len = 1,
  big.len,
  small.age,
  big.age,
  ref.age
)
}
\arguments{
\item{small.len}{Numeric scalar. Smallest length to be estimated (Usually based on observations for the year)}

\item{big.len}{Numeric scalar. Largest length to be estimated (Usually based on observations for the year)}

\item{small.age}{Numeric scalar.}

\item{big.age}{Numeric scalar. Maximum age needed in the key (Usually based on observations for the year)}

\item{ref.age}{Numeric scalar. Used in the GLM as the age against which all
the other ages are estimated. It is typically the most abundant age.}
}
\value{
List

\item{converg}{Indicating if the fit converged. }
\item{pred.p}{Matrix. Age-length Key. Rows are lengths, columns ages}
}
\description{
Function to compute proportions-at-age for each length using multinomial approach
From Chris Legault (which was based on code from Mike Bednarski)
}
