% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{get_sample_data}
\alias{get_sample_data}
\title{Pull a sample data set}
\usage{
get_sample_data(channel, species = 164744, species_itis = T, area = "all")
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}.
This object is used to communicate with the database engine. (see \code{\link[dbutils]{connect_to_database}})}

\item{species}{Numeric scalar. Species_itis or NESPP3 code. Default = 164744 (147, Haddock)}

\item{area}{Numeric vector. Statistical areas for which to pull data from. Default = "all"}

\item{species_itis.}{Boolean . TRUE indicates species code is species_itis, FALSE = nespp3}
}
\value{
A list of two objects

\item{landingsData}{Tibble containing landing data in format required for processing}
\item{lengthData}{Tibble containing landing data in format required for processing}
}
\description{
Data in the format needed for processing.
Landings are pulled from STOCKEFF database using the tables:
"MV_CF_Landings" for landings and
"MV_CF_LEN" for lengths
}
\details{
THIS IS A SAMPLE DATASET.

There are missing Areas in the database which wont be used.

There are QTR values coded as zero which wont be used.

All of the missing data needs to be dealt with prior to using the processing algorithm

To get access to the data base you will need credentials, permissions, and software. See \code{\link[dbutils]{connect_to_database}}
}
\examples{
\dontrun{
#Pull all Haddock data
channel <- dbutils::connect_to_database("serverName","userName")
data <- get_sample_data(channel,species = 164744, area="all")
}

}
