% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_data.r
\name{get_species_data}
\alias{get_species_data}
\title{Pull species data from cfdbs}
\usage{
get_species_data(
  channel,
  speciesRules = NULL,
  year = "all",
  outPath = here::here(),
  fileName = NULL
)
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}.
This object is used to communicate with the database engine. (see \code{\link[dbutils]{connect_to_database}})}

\item{speciesRules}{List. Containing species specific rules. Default = NULL (Fully automated).
Note:  Predefined \code{speciesRules} will be bundled with the package for select species}

\item{year}{Numeric vector. Containing the years to search over. Default = "all"}

\item{outPath}{Character string. Path to directory where data will be saved}

\item{fileName}{character string. File name of data to be saved}
}
\value{
List. Data is also written to disk.

\item{landins}{Data frame containing landings data by YEAR, QTR, NEGEAR, MARKET_CODE}
\item{lengths}{Data frame containing fish lengths by YEAR, QTR, NEGEAR, MARKET_CODE }
}
\description{
Pulls data from cfdbs via \code{\link{[comlandr]}}
Palmer method applied to assign missing attributes (NEGEAR, QTR, AREA) to landings records
}
