% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_object.r
\name{get_species_object}
\alias{get_species_object}
\title{Get species object}
\usage{
get_species_object(species_itis)
}
\arguments{
\item{species_itis}{Numeric. Itis code for species}
}
\description{
Lists set of rules required for catch at length aggregation
}
