% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketCodeLookupTable.R
\docType{data}
\name{marketCodeLookupTable}
\alias{marketCodeLookupTable}
\title{A lookup table for species market codes}
\format{
A dataframe with 7 columns
\describe{
  \item{COMMON_NAME}{Common name for species}
  \item{SCIENTIFIC_NAME}{scientific name for species}
  \item{SPECIES_ITIS}{define this}
  \item{NESPP4}{US North east 4 digit species code}
  \item{MARKET_DESC}{Description of the fish part}
  \item{MARKET_CODE}{2 digit market code}
  \item{MARKET_SCALE}{Ordinal scale for market code within species}
}
}
\source{
SPECIES_ITIS_NE table
}
\usage{
marketCodeLookupTable
}
\description{
Market codes are defined by the 4th digit of an NESPP4 code. These codes are mapped to descriptions.
This data set includes the codes and descriptions of market codes for a set of 10 species
}
\keyword{datasets}
