% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_length_by_year.R
\name{missing_length_by_year}
\alias{missing_length_by_year}
\title{Function to fill in missing sample lengths by year}
\usage{
missing_length_by_year(YEARData, targetYEAR, minNumSamples)
}
\description{
Obtain the YEAR length samples to be used for missing YEAR data
Assuption is the data has been aggregated to YEAR from QTR.
The closest year that contains length samples is chosen
}
\details{
Internal function
}
