% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscatch.R
\docType{package}
\name{mscatch}
\alias{mscatch}
\alias{mscatch-package}
\title{mscatch: landings data manipulation}
\description{
Tools to aid in extraction, manipulation of data from the cfdbs database
}
\details{
\itemize{
\item Pull spatial data from cfdbs and aggregate to stat area/EPU.
\item Methods for assigning trips to missing areas
\item Methods for assigning trips to EPUs when 10 min square is missing
\item Methods for aggregating landings based on length sample frequency (as for single species assessments)
\item Methods for distributing landings by length/weight (by using length-weight keys)
\item + more
}

To learn more about using \code{mscatch}, start with the vignette: \code{browseVignettes(package="mscatch")} or click the index below
}
