% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landings_by_gear.R
\name{plot_landings_by_gear}
\alias{plot_landings_by_gear}
\title{plot landings by gear}
\usage{
plot_landings_by_gear(species, data, threshold, outputPlots, outputDir,
  plotID)
}
\description{
Select the variable to plot.
Eventually have the option to print to file as diagnostics
}
