% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landings_data.R
\name{plot_landings_data}
\alias{plot_landings_data}
\title{plot landings data as checks to aggregation rules}
\usage{
plot_landings_data(species, data, dataType, threshold)
}
\description{
Select the variable to plot.
Eventually have the option to print to file as diagnostics
}
