% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_market_code_by_qtr.R
\name{plot_market_code_by_qtr}
\alias{plot_market_code_by_qtr}
\title{plots Number of length samples}
\usage{
plot_market_code_by_qtr(data, plotID, outputDir, outputPlots = T)
}
\arguments{
\item{data}{List. landings and length data aggregated to some level}

\item{plotID}{numeric scalar. number ID of plot}

\item{outputDir}{Character string. Path to output directory (png files saved here)}

\item{outputPlots}{Boolean. Should plots be created. T or F}
}
\value{
Nothing
}
\description{
A facet grid for each gear type by market code and quarter year
Eventually generalize to semester or annually
}
