% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_market_codes.R
\name{plot_market_codes}
\alias{plot_market_codes}
\title{plots landings by market code}
\usage{
plot_market_codes(marketData, plotID, outputDir, outputPlots)
}
\arguments{
\item{marketData}{data frame.}

\item{plotID}{numeric scalar. number ID of plot}

\item{outputDir}{Character string. Path to output directory (png files saved here)}

\item{outputPlots}{Boolean. Should plots be created. T or F}
}
\description{
Plots landings, number of samples for all market_codes found in landings data
}
