% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesLookupTable.R
\docType{data}
\name{speciesLookupTable}
\alias{speciesLookupTable}
\title{A lookup table for species codes}
\format{
A dataframe with 6 columns
\describe{
  \item{SCIENTIFIC_NAME}{scientific name for species}
  \item{SPECIES_ITIS}{define this}
  \item{COMMON_NAME.x}{Common name for species}
  \item{NESPP3}{US North east 3 digit species code for commercially landed species}
  \item{COMMON_NAME.x}{Common name for species}
  \item{SVSPP}{US North east 3 digit species code for survey species}
}
}
\usage{
speciesLookupTable
}
\description{
Species of interest along with species codes
}
\keyword{datasets}
