% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_early_years.R
\name{update_early_years}
\alias{update_early_years}
\title{Fill in missing length sample data for early years}
\usage{
update_early_years(QTRData, maxEarlyYear, iqtr, nLengthSamples)
}
\arguments{
\item{QTRData}{Tibble. (n x 8). Filtered landings data by NEGEAR, MARKET_CODE, YEARs > earlyYears}

\item{maxEarlyYear}{numeric scalar. The last year before length samplings began}

\item{iqtr}{numic scalar. QTR in which to pull length data from}

\item{nLengthSamples}{numeric scalar.The minimum number of length sample sizes required to avoid combination of data. Dfeault = 1}
}
\value{
\item{numSamples}{tibble (nx4). Columns =  YEAR, QTR, len_totalNumLen, len_numLengthSamples}
}
\description{
Data shows landings recorded prior to a time when length samples were taken.
eg. Landings in 1965 but length sampling began in 1969. Length samples are asscoiated with these earlier
landings.
}
