% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data_by_class.R
\name{aggregate_data_by_class}
\alias{aggregate_data_by_class}
\title{Aggregates rows of data by a given variable}
\usage{
aggregate_data_by_class(data, variable, classes, dataset)
}
\arguments{
\item{data}{tibble. Data set to aggreate}

\item{variable}{Character string. Name of the variable to aggregate over}

\item{classes}{Character vector. Two codes (from the variable) to aggregate}

\item{dataset}{Characterstring. Denoting the type of data set. "landings" or "lengths}
}
\description{
Collapses the data based on variables to collapse over
}
