% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gear.R
\name{aggregate_gear}
\alias{aggregate_gear}
\title{Aggregates landings and length data based on gear}
\usage{
aggregate_gear(
  data,
  recodeOtherGear,
  landingsThresholdGear,
  species_itis,
  outputDir,
  outputPlots
)
}
\arguments{
\item{data}{List. Landings data and length data}

\item{recodeOtherGear}{Numeric scalar. Arbitrary code to use for "other gears" that contribute little to landings}

\item{landingsThresholdGear}{Numeric scalar. (Proportion). Determines the proportion of landings that have to be satisfied before all other landings by gear are lumped into "other gear" category.
This is passed via the main function (\code{aggregate_landings})}
}
\value{
List

\item{landings}{same as input}
\item{lengthData}{Same as input}
}
\description{
Selects unique gear types that comprise a certain percentage of total landings. All landings by other gear types are deemed minor
 and thus aggregated into an "other gear" category. The percentage is an function argument
}
