% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_age_data.R
\name{combine_age_data}
\alias{combine_age_data}
\title{Combine age data from svdbs and cfdbs}
\usage{
combine_age_data(ageSurvey, ageComland, aggregate_to)
}
\arguments{
\item{ageSurvey}{Data frame}

\item{ageComland}{Data frame}

\item{aggregate_to}{Character string. Level of aggregation for all MARKET_CODES and gears ("QTR", "YEAR", "SEMESTER", MIX").
Default = YEAR.}
}
\value{
Data frame of combined length and ages across time

\item{}{}
\item{}{}
}
\description{
Combines the age data from the survey and commercial samples
}
