% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_length_distributions.R
\name{compare_length_distributions}
\alias{compare_length_distributions}
\title{Compare length distributions across categories}
\usage{
compare_length_distributions(data, variableToAggregate, groupBy, pValue,
  outputDir, logfile)
}
\arguments{
\item{data}{List. Landings and length data in form described below}

\item{pValue}{numeric scalar. Threshold pvalue for determining significance of ks test for length samples}

\item{outputDir}{Character string. Directory where output is saved}

\item{logfile}{Cahracter string. Name of logfile}

\item{variableToAggreagte}{character string. Denote the variable to aggregate based on length distribution}
}
\value{
Character vector
\item{codes}{Two market category codes not significantly different}
}
\description{
Compare length distributions across categories
}
