% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_length_weight.R
\name{fit_length_weight}
\alias{fit_length_weight}
\title{Fit length weight relationship to species data from SVDBS}
\usage{
fit_length_weight(
  lengthWeightData,
  speciesName,
  speciesRules,
  outputDir = NULL,
  logfile = NULL,
  suppressMessages = F
)
}
\arguments{
\item{lengthWeightData}{Data frame. length-weight pairs. Each row represents an individual fish}

\item{speciesName}{Character string. Common name for species}

\item{speciesRules}{List. Obtained from \code{get_species_object}}

\item{outputDir}{Character string. Path to output directory (Default = NULL, no output written)}

\item{logfile}{Character string. Specify the name for the log file generated describing all decisions made.
(Default = NULL, no output written)}

\item{suppressMessages}{Boolean. Suppress all messages}
}
\value{
List of model fit objects
\item{commonSlope}{\code{\link{lm}} object. Fit for single slope (beta)}
\item{seasonalSlope}{\code{\link{lm}} object. Fit for seasonal slopes}
}
\description{
Fits a length weight relationship for use in catch expansion
}
\section{Notes on model fitting}{


The Weight-Length relationship is defined as

\deqn{W_{ij} = \alpha L_{ij}^{\beta_j} e^{z_{ij}}}

where,

\eqn{W_{ij}} = Weight of fish \emph{i} in season \emph{j}, \eqn{i = 1, ..., n, j = 1, ... J}

\eqn{L_{ij}} = Length of fish \emph{i} in season \emph{j},

\eqn{z_{ij}} ~ \eqn{N(0,\sigma^2)},

and \eqn{\beta_j} is effect of season \emph{j}

On the more familiar log scale the model is

\deqn{log(W_{ij}) = log(\alpha) + \beta_j log(L_{ij}) + {z_{ij}}}

To test for a seasonal effect, we test the Null hypothesis:

 \deqn{{H_0}:  \beta_j = \beta}

 against the alternative,

 \deqn{{H_1}: \beta_j \neq \beta}

The test statistic is the standard F statistic

\deqn{F = \frac{(RSS_{H_0}-RSS_{H_1})/(J-1)}{RSS_{H_1}/(n-J-1)}}

which will have an F distribution with (J-1, n-J-1) degrees of freedom

where RSS= Residual Sum of Square s
}

