% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_data.r
\name{get_species_data}
\alias{get_species_data}
\title{Pull species data from cfdbs}
\usage{
get_species_data(
  species_itis = NULL,
  stock = NULL,
  outPath = here::here(),
  fileName = NULL
)
}
\arguments{
\item{species_itis}{Numeric. Itis code for species (Default = NA, Returns a NA object for user to define)}

\item{stock}{Character string. Define the stock associated with species of interest.
eg. for winter flounder stock = "GB" or "SNEMA". For black sea bass, "North" or "South"}

\item{outPath}{Character string. Path to directory where data will be saved}

\item{fileName}{character string. File name of data to be saved}
}
\value{
List. Data is also written to disk.

\item{landins}{Data frame containing landings data by YEAR, QTR, NEGEAR, MARKET_CODE}
\item{lengths}{Data frame containing fish lengths by YEAR, QTR, NEGEAR, MARKET_CODE }
}
\description{
Pulls data from cfdbs via \code{\link{[comlandr]}}
Palmer method applied to assign missing attributes (NEGEAR, QTR, AREA) to landings records
}
