% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_length_by_qtr.R
\name{missing_length_by_qtr}
\alias{missing_length_by_qtr}
\title{Recursive function to fill in missing sample lengths by quarter}
\usage{
missing_length_by_qtr(QTRData, missingQTRs, iyear)
}
\description{
Obtain the YEAR and QTR length samples to be used for missing YEAR/QTR data
}
