% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_length_by_year.R
\name{missing_length_by_year}
\alias{missing_length_by_year}
\title{Function to fill in missing sample lengths by year}
\usage{
missing_length_by_year(YEARData, targetYEAR, minNumSamples)
}
\arguments{
\item{YEARData}{dataframe. YEAR and number of length samples for the year}

\item{minNumSamples}{numeric scalar. Threshold for number of samples (This is set in main function and passed)}

\item{targetYear}{numeric scalar. Year with missing length samples}
}
\value{
\item{numSampels}{description}
}
\description{
Obtain the YEAR length samples to be used for missing YEAR data
Assuption is the data has been aggregated to YEAR from QTR.
The closest year that contains length samples is chosen
}
