% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscatch.R
\docType{package}
\name{mscatch}
\alias{mscatch}
\title{mscatch: Expanding total catch for species by length composition}
\description{
Tools to aid in aggregation of catch data and expanding catch data to weight by length composition
}
\details{
\itemize{
\item Pull spatial data from cfdbs and aggregate
\item Methods for aggregating landings based on length sample frequency
\item Methods for distributing landings by length/weight (by using length-weight keys)
}

To learn more about using \code{mscatch}, start with the vignette: \code{browseVignettes(package="mscatch")} or click the index below
}
