% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_length_histograms.R
\name{plot_length_histogram}
\alias{plot_length_histogram}
\title{plots length distributions}
\usage{
plot_length_histogram(lengthData, species_itis, outputDir)
}
\arguments{
\item{lengthData}{data frame.}

\item{species_itis}{Numeric scalar.  Species_itis code for species}
}
\description{
Creates a facet plot for each market category across all years for each gear type
}
