% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_aggregation.R
\name{test_aggregation}
\alias{test_aggregation}
\title{Test aggregation}
\usage{
test_aggregation(landingsThreshold = c(1000, 10))
}
\arguments{
\item{landingsThreshold}{vector (length 2). minimum landings (lbs) and minimum sample size to avoid combination of data}
}
\description{
This will all change so no point documenting it too much
}
