% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_aggregation.R
\name{test_aggregation}
\alias{test_aggregation}
\title{Test aggregation}
\usage{
test_aggregation(landingsData = sampleData_164744,
  lengthData = sampleLengths_164744, species_itis = 164744,
  landingsThresholdGear = 0.9, nLengthSamples = 1, pValue = 0.05,
  proportionMissing = 0.2, outputDir = here::here("output"),
  outputPlots = F, logfile = "logFile.txt")
}
\arguments{
\item{landingsData}{tidy data frame. Landings by YEAR,QTR,NEGEAR,MARKET_CODE,landings_land,landings_nn,len_totalNumLen,len_numLengthSampls}

\item{lengthData}{tidy data frame. Length data by YEAR,QTR,NEGEAR,MARKET_CODE, LENGTH, NUMLEN}

\item{species_itis}{numeric scalar. species_itis code relating to the landings and length data}

\item{landingsThresholdGear}{numeric scalar (proportion). Minimum proportion of cumulative landings to avoid aggregation of gear. Default = .9}

\item{nLengthSamples}{numeric scalar. The minimum number of length sample sizes required to avoid combination of data. Default = 1}

\item{pValue}{numeric scalar. Threshold pvalue for determining significance of ks test for length samples}

\item{proportionMissing}{numeric scalar. Proportion of missing samples allowed per YEAR for each MARKET_CODE/GEAR combination). Default = 0.2}

\item{outputDir}{Character string. Path to output directory (png files saved here)}

\item{outputPlots}{Boolean. Should plots be created. T or F (Default = F)}

\item{logFile}{character string. Specify the name for the log file generated describing all decisions made.}
}
\description{
This will all change so no point documenting it too much.
Eventually split each section into its own function for easier future development
All plotting functions will need to be finined/generalized
Also need to adapt code to allow user to process on a year by year basis
}
