% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_aggregation.R
\name{test_aggregation}
\alias{test_aggregation}
\title{Test aggregation}
\usage{
test_aggregation(landingsThresholdGear = 0.9, nLengthSamples = 1,
  outputDir = here::here("output"), outputPlots = F)
}
\arguments{
\item{landingsThresholdGear}{numeric scalar (proportion). Minimum proportion of cumulative landings to avoid aggregation of gear. Default = .9}

\item{nLengthSamples}{numeric scalar. The minimum number of length sample sizes required to avoid combination of data. Dfault = 1}

\item{outputDir}{Character string. Path to output directory (png files saved here)}

\item{outputPlots}{Boolean. Should plots be created. T or F (Default = F)}
}
\description{
This will all change so no point documenting it too much.
Eventually split each section into its own function for easier future development
}
