% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_aggregation.R
\name{test_aggregation}
\alias{test_aggregation}
\title{Test aggregation}
\usage{
test_aggregation(landingsThreshold = 0.9, nLengthSamples = 1)
}
\arguments{
\item{landingsThreshold}{numeric scalar (proportion). Minimum proportion of cumulative landings to avoid aggregation of gear. Default = .9}

\item{nLengthSamples}{numeric scalar. The minimum number of length sample sizes required to avoid combination of data. Dfault = 1}
}
\description{
This will all change so no point documenting it too much.
Eventually split each section into its own function for easier future development
}
