% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_pull.R
\name{test_data_pull}
\alias{test_data_pull}
\title{sample Data pull}
\usage{
test_data_pull(channel, species = 212)
}
\arguments{
\item{channel}{obtained from logging into the sole database.}

\item{species}{nespp3 code. default = 212 (atlantic mackerel)}
}
\description{
All missing areas need to dealt with prior to pulling species data.
Not the case in this sample. We aggregate all data as if from one EPU
}
