% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_early_years.R
\name{update_early_years}
\alias{update_early_years}
\title{Fill in missing length sample data for early years}
\usage{
update_early_years(QTRData, maxEarlyYear, iqtr)
}
\arguments{
\item{QTRData}{data}

\item{maxEarlyYear}{numeric scalar. The last year before length samplings began}

\item{iqtr}{numic scalar. QTR in which to pull length data from}
}
\value{
\item{numSamples}{description}
}
\description{
Data shows landings recorded prior to a time when length samples were taken.
eg. Landings in 1965 but length sampling began in 1969. Length samples are asscoiated with these earlier
landings.
}
