% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_length_samples.R
\name{update_length_samples}
\alias{update_length_samples}
\title{update missing length values}
\usage{
update_length_samples(data, missingRow, gearType, marketCode, numSamples)
}
\description{
For each YEAR/QRT/NEGEAR/MARKET_CODE with missing length samples we update with previous years data
}
\details{
internal
}
