% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_length_samples.R
\name{update_length_samples}
\alias{update_length_samples}
\title{update missing length values}
\usage{
update_length_samples(data, missingRow, gearType, marketCode, numSamples)
}
\arguments{
\item{data}{List. landings and lengthData}

\item{missingRow}{Dataframe. Specifies the YEAR and QTR where length samples are missing}

\item{gearType}{Character string. Name of NEGEAR}

\item{marketCode}{Character string. Name of MARKET_CODE}

\item{numSamples}{Tibble. Filtered landings data (YEAR, QTR, len_totalNumLen, len_numLengthSamples) by NEGEAR, MARKET_CODE}
}
\value{
List of landings and associated length samples .Same as input data
\item{landings}{Tibble (n x 8). Aggregated landings data. YEAR, QTR, NEGEAR, MARKET_CODE,landings_land (metric tonnes), landings_nn (# trips), len_totalNumLen (# fish lengths), len_numLengthSamples (# independent samples) }
\item{lengthData}{Tibble (m x 8 ). Aggregated length data. YEAR, QTR, NEGEAR, MARKET_CODE, LENGTH (length of fish), NUMLEN (# fish at LENGTH)}
}
\description{
For each YEAR/QRT/NEGEAR/MARKET_CODE with missing length samples we update with previous years data.
No aggregation is undertaken. Simply populating empty cells or adding length data to previously omitted cells.
}
